# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RedEvent API",
        "description": "Servicio RestFul para el aplicativo m\u00f3vil RedEvent.",
        "contact": {
            "email": "erick.benites@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "eventos-ebenites.c9users.io",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "List events",
                "description": "",
                "responses": {
                    "200": {
                        "description": "A list with events"
                    }
                }
            }
        },
        "/api/topics": {
            "get": {
                "tags": [
                    "Topics"
                ],
                "summary": "List topics",
                "description": "",
                "responses": {
                    "200": {
                        "description": "A list with topics"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login or Register",
                "description": "",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A user info"
                    },
                    "401": {
                        "description": "Password Invalid or Blocked account"
                    }
                }
            }
        },
        "/api/glogin": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login or Register with Google",
                "description": "",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "googleid",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "fullname",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "photo",
                        "in": "formData",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A user info"
                    },
                    "401": {
                        "description": "GoogleID Invalid or Blocked account"
                    }
                }
            }
        },
        "/api/users/{id}/topics": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List topics by user id",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with topics by user id"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update topics preferences",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {

                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user profile",
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {

                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by id",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A user info"
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
        "/api/users/{id}/events": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List events by user id",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with events by user id"
                    }
                }
            }
        },
        "/api/users/{id}/myevents": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List events registered",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with events by user id"
                    }
                }
            }
        },
        "/api/users/{userid}/events/{eventid}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Checking into envent",
                "description": "",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "eventid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register into envent",
                "description": "",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "eventid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message"
                    }
                }
            }
        },
        "/api/users/{userid}/events/{eventid}/{rating}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Rating envent",
                "description": "",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "eventid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message"
                    }
                }
            }
        }
    },
    "definitions": {

    }
}